Moved Card and related enums into it's own package
