+++ src/main/java/Logik/Cabal.java
@@ -2,7 +2,7 @@ package Logik;
 
 import Logik.cabal.internals.CardStack;
 import Logik.cabal.I_CabalModel;
-import Logik.cabal.internals.I_CardModel;
+import Logik.cabal.internals.card.I_CardModel;
 
 import java.util.List;
