+++ src/main/java/Logik/cabal/I_CabalModel.java
@@ -1,6 +1,6 @@
 package Logik.cabal;
 
-import Logik.cabal.internals.I_CardModel;
+import Logik.cabal.internals.card.I_CardModel;
 
 public interface I_CabalModel {
