+++ src/main/java/Logik/cabal/internals/CardStack.java
@@ -1,5 +1,7 @@
 package Logik.cabal.internals;
 
+import Logik.cabal.internals.card.E_CardSuit;
+import Logik.cabal.internals.card.I_CardModel;
 import Logik.error.IllegalMoveException;
 
 import java.util.ArrayList;
