+++ src/main/java/Logik/cabal/internals/card/E_CardSuit.java
@@ -1,4 +1,4 @@
-package Logik.cabal.internals;
+package Logik.cabal.internals.card;
 
 /**
  * This is the enumerator for the suit of the card.
