+++ src/test/java/Logik/cabal/Cabal_madsTest.java
@@ -1,7 +1,10 @@
 package Logik.cabal;
 
-import Logik.cabal.Cabal_mads;
 import Logik.cabal.internals.*;
+import Logik.cabal.internals.card.Card;
+import Logik.cabal.internals.card.E_CardRank;
+import Logik.cabal.internals.card.E_CardSuit;
+import Logik.cabal.internals.card.I_CardModel;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
