+++ src/test/java/Logik/cabal/internals/CardStackTest.java
@@ -1,6 +1,9 @@
 package Logik.cabal.internals;
 
-import Logik.cabal.internals.*;
+import Logik.cabal.internals.card.Card;
+import Logik.cabal.internals.card.E_CardRank;
+import Logik.cabal.internals.card.E_CardSuit;
+import Logik.cabal.internals.card.I_CardModel;
 import Logik.error.IllegalMoveException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
