+++ src/test/java/Logik/cabal/internals/CardTest.java
@@ -1,9 +1,9 @@
 package Logik.cabal.internals;
 
-import Logik.cabal.internals.Card;
-import Logik.cabal.internals.E_CardRank;
-import Logik.cabal.internals.E_CardSuit;
-import Logik.cabal.internals.I_CardModel;
+import Logik.cabal.internals.card.Card;
+import Logik.cabal.internals.card.E_CardRank;
+import Logik.cabal.internals.card.E_CardSuit;
+import Logik.cabal.internals.card.I_CardModel;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
